requirements.txt
discord
discord.py
aiohttp
websockets
async-timeout
attrs
certifi
chardet
finnhub-python
idna
multidict
requests
typing-extensions
yarl
python-dotenv
